requirements.txt
torch
torchvision
tqdm
Pillow
editdistance
matplotlib